Apply filter state and sidebar to schedule apps.
