+++ test/integration/patients/sidebar/filter-sidebar.js
@@ -591,5 +591,736 @@ context('filter sidebar', function() {
       .find('.fa-square')
       .should('have.length', 0);
   });
+
+  specify('schedule filtering', function() {
+    localStorage.setItem(`schedule_11111_11111-${ STATE_VERSION }`, JSON.stringify({
+      filters: {
+        insurance: 'Medicare',
+      },
+      states: ['22222', '33333'],
+      flowStates: ['22222', '33333'],
+    }));
+
+    cy
+      .routeDirectories(fx => {
+        fx.data = [
+          {
+            attributes: {
+              name: 'Team',
+              slug: 'team',
+              value: [
+                'Coordinator',
+                'Nurse',
+              ],
+            },
+          },
+          {
+            attributes: {
+              name: 'Insurance Plans',
+              slug: 'insurance',
+              value: [
+                'BCBS PPO 100',
+                'Medicare',
+              ],
+            },
+          },
+          {
+            attributes: {
+              name: 'ACO',
+              slug: 'aco',
+              value: [
+                'Basic',
+                'Premier',
+              ],
+            },
+          },
+        ];
+
+        return fx;
+      })
+      .routeSettings(fx => {
+        fx.data.push({ id: 'custom_filters', attributes: { value: ['team', 'insurance'] } });
+
+        return fx;
+      })
+      .routeActions()
+      .visit('/schedule')
+      .wait('@routeActions')
+      .itsUrl()
+      .its('search')
+      .should('contain', 'filter[clinician]=11111')
+      .should('contain', 'filter[@insurance]=Medicare')
+      .should('contain', 'filter[state]=22222,33333');
+
+    cy
+      .get('.list-page__filters')
+      .find('[data-filters-region]')
+      .find('button')
+      .should('contain', '1');
+
+    cy
+      .get('.list-page__filters')
+      .find('[data-filters-region]')
+      .find('button')
+      .click();
+
+    cy
+      .get('.app-frame__sidebar .sidebar')
+      .as('filtersSidebar')
+      .find('.sidebar__heading')
+      .should('contain', '1');
+
+    cy
+      .get('@filtersSidebar')
+      .find('.js-clear-filters')
+      .should('not.be.disabled');
+
+    cy
+      .get('@filtersSidebar')
+      .find('[data-filter-button]')
+      .should('have.length', 2)
+      .first()
+      .get('.sidebar__label')
+      .should('contain', 'Insurance Plans')
+      .get('[data-filter-button')
+      .should('contain', 'Medicare');
+
+    cy
+      .get('@filtersSidebar')
+      .find('[data-filter-button]')
+      .eq(1)
+      .get('.sidebar__label')
+      .should('contain', 'Team')
+      .get('[data-filter-button')
+      .should('contain', 'All');
+
+    cy
+      .get('@filtersSidebar')
+      .find('[data-filter-button]')
+      .first()
+      .click();
+
+    cy
+      .get('.picklist')
+      .find('.js-input')
+      .should('have.attr', 'placeholder', 'Insurance Plans...');
+
+    cy
+      .get('.picklist__item')
+      .contains('All')
+      .click()
+      .then(() => {
+        const storage = JSON.parse(localStorage.getItem(`schedule_11111_11111-${ STATE_VERSION }`));
+
+        expect(storage.filters.insurance).to.be.null;
+      })
+      .wait('@routeActions')
+      .itsUrl()
+      .its('search')
+      .should('contain', 'filter[clinician]=11111')
+      .should('not.contain', 'filter[@insurance]')
+      .should('contain', 'filter[state]=22222,33333');
+
+    cy
+      .get('.list-page__filters')
+      .find('[data-filters-region]')
+      .find('button')
+      .should('not.contain', '1');
+
+    cy
+      .get('@filtersSidebar')
+      .find('.sidebar__heading')
+      .should('not.contain', '1');
+
+    cy
+      .get('@filtersSidebar')
+      .find('.js-clear-filters')
+      .should('be.disabled');
+
+    cy
+      .get('@filtersSidebar')
+      .find('[data-filter-button]')
+      .first()
+      .click();
+
+    cy
+      .get('.picklist__item')
+      .contains('BCBS PPO 100')
+      .click()
+      .then(() => {
+        const storage = JSON.parse(localStorage.getItem(`schedule_11111_11111-${ STATE_VERSION }`));
+
+        expect(storage.filters.insurance).to.equal('BCBS PPO 100');
+      })
+      .wait('@routeActions')
+      .itsUrl()
+      .its('search')
+      .should('contain', 'filter[clinician]=11111')
+      .should('contain', 'filter[@insurance]=BCBS PPO 100')
+      .should('contain', 'filter[state]=22222,33333');
+
+    cy
+      .get('.list-page__filters')
+      .find('[data-filters-region]')
+      .find('button')
+      .should('contain', '1');
+
+    cy
+      .get('@filtersSidebar')
+      .find('.sidebar__heading')
+      .should('contain', '1');
+
+    cy
+      .get('@filtersSidebar')
+      .find('.js-clear-filters')
+      .click()
+      .then(() => {
+        const storage = JSON.parse(localStorage.getItem(`schedule_11111_11111-${ STATE_VERSION }`));
+
+        expect(storage.filters.insurance).to.be.undefined;
+      })
+      .wait('@routeActions')
+      .itsUrl()
+      .its('search')
+      .should('contain', 'filter[clinician]=11111')
+      .should('not.contain', 'filter[@insurance]')
+      .should('contain', 'filter[state]=22222,33333');
+
+    cy
+      .get('.list-page__filters')
+      .find('[data-filters-region]')
+      .find('button')
+      .should('not.contain', '1');
+
+    cy
+      .get('@filtersSidebar')
+      .find('.sidebar__heading')
+      .should('not.contain', '1');
+
+    cy
+      .get('@filtersSidebar')
+      .find('[data-states-filters-region]')
+      .get('.sidebar__heading')
+      .should('contain', 'States');
+
+    cy
+      .get('@filtersSidebar')
+      .find('[data-states-filters-region]')
+      .should('contain', 'To Do')
+      .should('contain', 'In Progress')
+      .should('contain', 'Done')
+      .should('contain', 'Unable to Complete')
+      .should('contain', 'THMG Transfered');
+
+    cy
+      .get('@filtersSidebar')
+      .find('[data-states-filters-region]')
+      .find('.fa-square-check')
+      .should('have.length', 2);
+
+    cy
+      .get('@filtersSidebar')
+      .find('[data-states-filters-region]')
+      .find('[data-check-region]')
+      .first()
+      .click()
+      .then(() => {
+        const storage = JSON.parse(localStorage.getItem(`schedule_11111_11111-${ STATE_VERSION }`));
+
+        expect(storage.states).to.deep.equal(['33333']);
+      })
+      .wait('@routeActions')
+      .itsUrl()
+      .its('search')
+      .should('contain', 'filter[clinician]=11111')
+      .should('contain', 'filter[state]=33333')
+      .should('not.contain', 'filter[state]=22222');
+
+    cy
+      .get('.list-page__filters')
+      .find('[data-filters-region]')
+      .find('button')
+      .should('contain', '1');
+
+    cy
+      .get('@filtersSidebar')
+      .find('.sidebar__heading')
+      .should('contain', '1');
+
+    cy
+      .get('@filtersSidebar')
+      .find('[data-states-filters-region]')
+      .find('[data-check-region]')
+      .eq(1)
+      .click()
+      .then(() => {
+        const storage = JSON.parse(localStorage.getItem(`schedule_11111_11111-${ STATE_VERSION }`));
+
+        expect(storage.states).to.deep.equal([]);
+      })
+      .wait('@routeActions')
+      .itsUrl()
+      .its('search')
+      .should('contain', 'filter[clinician]=11111')
+      .should('contain', `filter[state]=${ NIL_UUID }`);
+
+    cy
+      .get('.list-page__filters')
+      .find('[data-filters-region]')
+      .find('button')
+      .should('contain', '1');
+
+    cy
+      .get('@filtersSidebar')
+      .find('.sidebar__heading')
+      .should('contain', '1');
+
+    cy
+      .get('@filtersSidebar')
+      .find('[data-filter-button]')
+      .first()
+      .click();
+
+    cy
+      .get('.picklist__item')
+      .contains('BCBS PPO 100')
+      .click()
+      .wait('@routeActions');
+
+    cy
+      .get('.list-page__filters')
+      .find('[data-filters-region]')
+      .find('button')
+      .should('contain', '2');
+
+    cy
+      .get('@filtersSidebar')
+      .find('.sidebar__heading')
+      .should('contain', '2');
+
+    cy
+      .get('@filtersSidebar')
+      .find('.js-clear-filters')
+      .click()
+      .then(() => {
+        const storage = JSON.parse(localStorage.getItem(`schedule_11111_11111-${ STATE_VERSION }`));
+
+        expect(storage.states).to.deep.equal(['22222', '33333']);
+      })
+      .wait('@routeActions')
+      .itsUrl()
+      .its('search')
+      .should('contain', 'filter[clinician]=11111')
+      .should('contain', 'filter[state]=22222,33333')
+      .should('not.contain', 'filter[@insurance]');
+
+    cy
+      .get('.list-page__filters')
+      .find('[data-filters-region]')
+      .find('button')
+      .should('not.contain', '2');
+
+    cy
+      .get('@filtersSidebar')
+      .find('.sidebar__heading')
+      .should('not.contain', '2');
+
+    cy
+      .get('@filtersSidebar')
+      .find('.js-close')
+      .click();
+
+    cy
+      .get('@filtersSidebar')
+      .should('not.exist');
+
+    cy
+      .get('.list-page__filters')
+      .find('[data-filters-region]')
+      .find('button')
+      .click();
+
+    cy
+      .get('[data-select-all-region]')
+      .click();
+
+    cy
+      .get('@filtersSidebar')
+      .find('.js-close')
+      .click();
+
+    cy
+      .get('@filtersSidebar')
+      .should('not.exist');
+
+    cy
+      .get('[data-filters-region]')
+      .find('.js-cancel')
+      .click();
+
+    cy
+      .get('.list-page__filters')
+      .find('[data-filters-region]')
+      .find('button')
+      .click();
+
+    cy
+      .get('@filtersSidebar')
+      .should('exist');
+  });
+
+  specify('reduced schedule filtering', function() {
+    localStorage.setItem(`reduced-schedule_11111_11111-${ STATE_VERSION }`, JSON.stringify({
+      filters: {
+        insurance: 'Medicare',
+      },
+      states: ['22222', '33333'],
+      flowStates: ['22222', '33333'],
+    }));
+
+    cy
+      .routesForPatientDashboard()
+      .routeCurrentClinician(fx => {
+        fx.data.id = '11111';
+        fx.data.attributes.enabled = true;
+        fx.data.relationships.role.data.id = '44444';
+        return fx;
+      })
+      .routeDirectories(fx => {
+        fx.data = [
+          {
+            attributes: {
+              name: 'Team',
+              slug: 'team',
+              value: [
+                'Coordinator',
+                'Nurse',
+              ],
+            },
+          },
+          {
+            attributes: {
+              name: 'Insurance Plans',
+              slug: 'insurance',
+              value: [
+                'BCBS PPO 100',
+                'Medicare',
+              ],
+            },
+          },
+          {
+            attributes: {
+              name: 'ACO',
+              slug: 'aco',
+              value: [
+                'Basic',
+                'Premier',
+              ],
+            },
+          },
+        ];
+
+        return fx;
+      })
+      .routeSettings(fx => {
+        fx.data.push({ id: 'custom_filters', attributes: { value: ['team', 'insurance'] } });
+
+        return fx;
+      })
+      .routeActions()
+      .routeAction()
+      .routePatientByAction()
+      .routeFormByAction()
+      .visit('/')
+      .wait('@routeActions')
+      .itsUrl()
+      .its('search')
+      .should('contain', 'filter[clinician]=11111')
+      .should('contain', 'filter[@insurance]=Medicare')
+      .should('contain', 'filter[state]=22222,33333');
+
+    cy
+      .get('.list-page__filters')
+      .find('[data-filters-region]')
+      .find('button')
+      .should('contain', '1');
+
+    cy
+      .get('.list-page__filters')
+      .find('[data-filters-region]')
+      .find('button')
+      .click();
+
+    cy
+      .get('.app-frame__sidebar .sidebar')
+      .as('filtersSidebar')
+      .find('.sidebar__heading')
+      .should('contain', '1');
+
+    cy
+      .get('@filtersSidebar')
+      .find('.js-clear-filters')
+      .should('not.be.disabled');
+
+    cy
+      .get('@filtersSidebar')
+      .find('[data-filter-button]')
+      .should('have.length', 2)
+      .first()
+      .get('.sidebar__label')
+      .should('contain', 'Insurance Plans')
+      .get('[data-filter-button')
+      .should('contain', 'Medicare');
+
+    cy
+      .get('@filtersSidebar')
+      .find('[data-filter-button]')
+      .eq(1)
+      .get('.sidebar__label')
+      .should('contain', 'Team')
+      .get('[data-filter-button')
+      .should('contain', 'All');
+
+    cy
+      .get('@filtersSidebar')
+      .find('[data-filter-button]')
+      .first()
+      .click();
+
+    cy
+      .get('.picklist')
+      .find('.js-input')
+      .should('have.attr', 'placeholder', 'Insurance Plans...');
+
+    cy
+      .get('.picklist__item')
+      .contains('All')
+      .click()
+      .then(() => {
+        const storage = JSON.parse(localStorage.getItem(`reduced-schedule_11111_11111-${ STATE_VERSION }`));
+
+        expect(storage.filters.insurance).to.be.null;
+      })
+      .wait('@routeActions')
+      .itsUrl()
+      .its('search')
+      .should('contain', 'filter[clinician]=11111')
+      .should('not.contain', 'filter[@insurance]')
+      .should('contain', 'filter[state]=22222,33333');
+
+    cy
+      .get('.list-page__filters')
+      .find('[data-filters-region]')
+      .find('button')
+      .should('not.contain', '1');
+
+    cy
+      .get('@filtersSidebar')
+      .find('.sidebar__heading')
+      .should('not.contain', '1');
+
+    cy
+      .get('@filtersSidebar')
+      .find('.js-clear-filters')
+      .should('be.disabled');
+
+    cy
+      .get('@filtersSidebar')
+      .find('[data-filter-button]')
+      .first()
+      .click();
+
+    cy
+      .get('.picklist__item')
+      .contains('BCBS PPO 100')
+      .click()
+      .then(() => {
+        const storage = JSON.parse(localStorage.getItem(`reduced-schedule_11111_11111-${ STATE_VERSION }`));
+
+        expect(storage.filters.insurance).to.equal('BCBS PPO 100');
+      })
+      .wait('@routeActions')
+      .itsUrl()
+      .its('search')
+      .should('contain', 'filter[clinician]=11111')
+      .should('contain', 'filter[@insurance]=BCBS PPO 100')
+      .should('contain', 'filter[state]=22222,33333');
+
+    cy
+      .get('.list-page__filters')
+      .find('[data-filters-region]')
+      .find('button')
+      .should('contain', '1');
+
+    cy
+      .get('@filtersSidebar')
+      .find('.sidebar__heading')
+      .should('contain', '1');
+
+    cy
+      .get('@filtersSidebar')
+      .find('.js-clear-filters')
+      .click()
+      .then(() => {
+        const storage = JSON.parse(localStorage.getItem(`reduced-schedule_11111_11111-${ STATE_VERSION }`));
+
+        expect(storage.filters.insurance).to.be.undefined;
+      })
+      .wait('@routeActions')
+      .itsUrl()
+      .its('search')
+      .should('contain', 'filter[clinician]=11111')
+      .should('not.contain', 'filter[@insurance]')
+      .should('contain', 'filter[state]=22222,33333');
+
+    cy
+      .get('.list-page__filters')
+      .find('[data-filters-region]')
+      .find('button')
+      .should('not.contain', '1')
+      .click();
+
+    cy
+      .get('@filtersSidebar')
+      .find('.sidebar__heading')
+      .should('not.contain', '1');
+
+    cy
+      .get('@filtersSidebar')
+      .find('[data-states-filters-region]')
+      .get('.sidebar__heading')
+      .should('contain', 'States');
+
+    cy
+      .get('@filtersSidebar')
+      .find('[data-states-filters-region]')
+      .should('contain', 'To Do')
+      .should('contain', 'In Progress')
+      .should('contain', 'Done')
+      .should('contain', 'Unable to Complete')
+      .should('contain', 'THMG Transfered');
+
+    cy
+      .get('@filtersSidebar')
+      .find('[data-states-filters-region]')
+      .find('.fa-square-check')
+      .should('have.length', 2);
+
+    cy
+      .get('@filtersSidebar')
+      .find('[data-states-filters-region]')
+      .find('[data-check-region]')
+      .first()
+      .click()
+      .then(() => {
+        const storage = JSON.parse(localStorage.getItem(`reduced-schedule_11111_11111-${ STATE_VERSION }`));
+
+        expect(storage.states).to.deep.equal(['33333']);
+      })
+      .wait('@routeActions')
+      .itsUrl()
+      .its('search')
+      .should('contain', 'filter[clinician]=11111')
+      .should('contain', 'filter[state]=33333')
+      .should('not.contain', 'filter[state]=22222');
+
+    cy
+      .get('.list-page__filters')
+      .find('[data-filters-region]')
+      .find('button')
+      .should('contain', '1');
+
+    cy
+      .get('@filtersSidebar')
+      .find('.sidebar__heading')
+      .should('contain', '1');
+
+    cy
+      .get('@filtersSidebar')
+      .find('[data-states-filters-region]')
+      .find('[data-check-region]')
+      .eq(1)
+      .click()
+      .then(() => {
+        const storage = JSON.parse(localStorage.getItem(`reduced-schedule_11111_11111-${ STATE_VERSION }`));
+
+        expect(storage.states).to.deep.equal([]);
+      })
+      .wait('@routeActions')
+      .itsUrl()
+      .its('search')
+      .should('contain', 'filter[clinician]=11111')
+      .should('contain', `filter[state]=${ NIL_UUID }`);
+
+    cy
+      .get('.list-page__filters')
+      .find('[data-filters-region]')
+      .find('button')
+      .should('contain', '1');
+
+    cy
+      .get('@filtersSidebar')
+      .find('.sidebar__heading')
+      .should('contain', '1');
+
+    cy
+      .get('@filtersSidebar')
+      .find('[data-filter-button]')
+      .first()
+      .click();
+
+    cy
+      .get('.picklist__item')
+      .contains('BCBS PPO 100')
+      .click()
+      .wait('@routeActions');
+
+    cy
+      .get('.list-page__filters')
+      .find('[data-filters-region]')
+      .find('button')
+      .should('contain', '2');
+
+    cy
+      .get('@filtersSidebar')
+      .find('.sidebar__heading')
+      .should('contain', '2');
+
+    cy
+      .get('@filtersSidebar')
+      .find('.js-clear-filters')
+      .click()
+      .then(() => {
+        const storage = JSON.parse(localStorage.getItem(`reduced-schedule_11111_11111-${ STATE_VERSION }`));
+
+        expect(storage.states).to.deep.equal(['22222', '33333']);
+      })
+      .wait('@routeActions')
+      .itsUrl()
+      .its('search')
+      .should('contain', 'filter[clinician]=11111')
+      .should('contain', 'filter[state]=22222,33333')
+      .should('not.contain', 'filter[@insurance]');
+
+    cy
+      .get('.list-page__filters')
+      .find('[data-filters-region]')
+      .find('button')
+      .should('not.contain', '2');
+
+    cy
+      .get('@filtersSidebar')
+      .find('.sidebar__heading')
+      .should('not.contain', '2');
+
+    cy
+      .get('@filtersSidebar')
+      .find('.js-close')
+      .click();
+
+    cy
+      .get('@filtersSidebar')
+      .should('not.exist');
+  });
 });
